+++ cmake/common.cmake
@@ -73,7 +73,11 @@ set(CCF_GENERATED_DIR ${CMAKE_CURRENT_BINARY_DIR}/generated)
 include_directories(${CCF_DIR}/include)
 include_directories(${CCF_DIR}/src)
 
-include_directories(SYSTEM ${CCF_DIR}/3rdparty)
+set(CCF_3RD_PARTY_EXPORTED_DIR "${CCF_DIR}/3rdparty/exported")
+set(CCF_3RD_PARTY_INTERNAL_DIR "${CCF_DIR}/3rdparty/internal")
+
+include_directories(SYSTEM ${CCF_3RD_PARTY_EXPORTED_DIR})
+include_directories(SYSTEM ${CCF_3RD_PARTY_INTERNAL_DIR})
 
 find_package(MbedTLS REQUIRED)
 
@@ -148,7 +152,7 @@ else()
 endif()
 
 # Lua module
-set(LUA_DIR ${CCF_DIR}/3rdparty/lua)
+set(LUA_DIR ${CCF_3RD_PARTY_INTERNAL_DIR}/lua)
 set(LUA_SOURCES
     ${LUA_DIR}/lapi.c
     ${LUA_DIR}/lauxlib.c
@@ -180,8 +184,16 @@ set(LUA_SOURCES
 )
 
 set(HTTP_PARSER_SOURCES
-    ${CCF_DIR}/3rdparty/llhttp/api.c ${CCF_DIR}/3rdparty/llhttp/http.c
-    ${CCF_DIR}/3rdparty/llhttp/llhttp.c
+    ${CCF_3RD_PARTY_EXPORTED_DIR}/llhttp/api.c
+    ${CCF_3RD_PARTY_EXPORTED_DIR}/llhttp/http.c
+    ${CCF_3RD_PARTY_EXPORTED_DIR}/llhttp/llhttp.c
+)
+
+set(CCF_ENDPOINTS_SOURCES
+    ${CCF_DIR}/src/endpoints/endpoint.cpp
+    ${CCF_DIR}/src/endpoints/endpoint_registry.cpp
+    ${CCF_DIR}/src/endpoints/base_endpoint_registry.cpp
+    ${CCF_DIR}/src/endpoints/common_endpoint_registry.cpp
 )
 
 set(CCF_ENDPOINTS_SOURCES
@@ -204,7 +216,9 @@ include(${CCF_DIR}/cmake/sss.cmake)
 function(add_unit_test name)
   add_executable(${name} ${CCF_DIR}/src/enclave/thread_local.cpp ${ARGN})
   target_compile_options(${name} PRIVATE ${COMPILE_LIBCXX})
-  target_include_directories(${name} PRIVATE src ${CCFCRYPTO_INC})
+  target_include_directories(
+    ${name} PRIVATE src ${CCFCRYPTO_INC} ${CCF_DIR}/3rdparty/test
+  )
   enable_coverage(${name})
   target_link_libraries(
     ${name} PRIVATE ${LINK_LIBCXX} ccfcrypto.host openenclave::oehost
@@ -267,7 +281,7 @@ if("virtual" IN_LIST COMPILE_TARGETS)
     # Remove the following two lines once we upgrade to snmalloc 0.5.4
     set(CMAKE_POLICY_DEFAULT_CMP0077 NEW)
     set(USE_POSIX_COMMIT_CHECKS off)
-    add_subdirectory(3rdparty/snmalloc EXCLUDE_FROM_ALL)
+    add_subdirectory(3rdparty/exported/snmalloc EXCLUDE_FROM_ALL)
     set(SNMALLOC_LIB snmalloc_lib)
     set(SNMALLOC_CPP src/enclave/snmalloc.cpp)
   endif()
@@ -636,7 +650,7 @@ function(add_picobench name)
   )
 
   # -Wall -Werror catches a number of warnings in picobench
-  target_include_directories(${name} SYSTEM PRIVATE 3rdparty)
+  target_include_directories(${name} SYSTEM PRIVATE 3rdparty/test)
 
   add_test(
     NAME ${name}