+++ src/node/backup_signatures.h
@@ -5,7 +5,6 @@
 #include "node_signature.h"
 #include "service_map.h"
 
-#include <msgpack/msgpack.hpp>
 #include <string>
 #include <vector>
 
@@ -18,8 +17,6 @@ namespace ccf
     crypto::Sha256Hash root;
     std::vector<NodeSignature> signatures;
 
-    MSGPACK_DEFINE(view, seqno, root, signatures);
-
     BackupSignatures() = default;
 
     BackupSignatures(
@@ -30,7 +27,7 @@ namespace ccf
     {}
   };
   DECLARE_JSON_TYPE(BackupSignatures);
-  DECLARE_JSON_REQUIRED_FIELDS(BackupSignatures, view, seqno, root, signatures)
+  DECLARE_JSON_REQUIRED_FIELDS(BackupSignatures, view, seqno, root, signatures);
 
   // Always recorded at key 0
   using BackupSignaturesMap = ServiceMap<size_t, BackupSignatures>;